Store file length in long variables
Variables holding a file length are now longs to prevent the risk of an
overrun on very large files.

+++ src/main/java/net/kemitix/kxssh/StatusProvider.java
@@ -4,7 +4,7 @@ public interface StatusProvider {
 
     void setStatusListener(StatusListener statusListener);
 
-    void updateProgress(int progress, int total);
+    void updateProgress(long progress, long total);
 
     void updateStatus(SshStatus status);
